requirements.txt
streamlit
langchain
langchain-openai
langchain-community
pypdf
requests
python-dotenv
beautifulsoup4
lxml